+++ Poseiden-skeleton/src/main/java/com/nnk/springboot/service/impl/BidListServiceImpl.java
@@ -1,9 +1,10 @@
-package com.nnk.springboot.service;
+package com.nnk.springboot.service.impl;
 
 import com.nnk.springboot.NotFoundException;
 import com.nnk.springboot.domain.BidList;
 import com.nnk.springboot.dto.BidListDto;
 import com.nnk.springboot.repositories.BidListRepository;
+import com.nnk.springboot.service.BidListService;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
